refactor [dev-9730] Chart Reducer
