+++ packages/chart/src/ConfigContext.tsx
@@ -1,5 +1,9 @@
-import { createContext } from 'react'
+import { createContext, Dispatch } from 'react'
 import { ChartContext } from './types/ChartContext'
+
+import ChartActions from './store/chart.actions'
+
+export const ChartDispatchContext = createContext<Dispatch<ChartActions>>(() => {})
 
 const ConfigContext = createContext({} as ChartContext)
