+++ packages/chart/src/components/Sparkline/components/SparkLine.tsx
@@ -9,6 +9,8 @@ import { MarkerArrow } from '@visx/marker'
 import ErrorBoundary from '@cdc/core/components/ErrorBoundary'
 import useReduceData from '../../../hooks/useReduceData'
 import ConfigContext from '../../../ConfigContext'
+import { getYAxisData, getXAxisData } from '../../../helpers/getAxisData'
+
 import './../index.scss'
 
 type SparkLineProps = {
@@ -18,7 +20,16 @@ type SparkLineProps = {
 
 const SparkLine: React.FC<SparkLineProps> = props => {
   const { width: parentWidth, height: parentHeight } = props
-  const { transformedData: data, config, parseDate, formatDate, seriesHighlight, formatNumber, colorScale, handleChartAriaLabels } = useContext(ConfigContext)
+  const {
+    transformedData: data,
+    config,
+    parseDate,
+    formatDate,
+    seriesHighlight,
+    formatNumber,
+    colorScale,
+    handleChartAriaLabels
+  } = useContext(ConfigContext)
   let width = Number(parentWidth)
   const { minValue, maxValue } = useReduceData(config, data, ConfigContext)
 
@@ -27,9 +38,6 @@ const SparkLine: React.FC<SparkLineProps> = props => {
 
   const xMax = width - config.runtime.yAxis.size
   const yMax = height - margin.top - 20
-
-  const getXAxisData = d => (config.runtime.xAxis.type === 'date' ? parseDate(d[config.runtime.originalXAxis.dataKey]).getTime() : d[config.runtime.originalXAxis.dataKey])
-  const getYAxisData = (d, seriesKey) => d[seriesKey]
 
   let xScale
   let yScale
@@ -53,7 +61,7 @@ const SparkLine: React.FC<SparkLineProps> = props => {
       max += paddingValue
     }
 
-    let xAxisDataMapped = data.map(d => getXAxisData(d))
+    let xAxisDataMapped = data.map(d => getXAxisData(d, config, parseDate))
 
     if (config.runtime.horizontal) {
       xScale = scaleLinear({
@@ -61,7 +69,10 @@ const SparkLine: React.FC<SparkLineProps> = props => {
         range: [0, xMax]
       })
 
-      yScale = config.runtime.xAxis.type === 'date' ? scaleLinear({ domain: [Math.min(...xAxisDataMapped), Math.max(...xAxisDataMapped)] }) : scalePoint({ domain: xAxisDataMapped, padding: 0.5 })
+      yScale =
+        config.runtime.xAxis.type === 'date'
+          ? scaleLinear({ domain: [Math.min(...xAxisDataMapped), Math.max(...xAxisDataMapped)] })
+          : scalePoint({ domain: xAxisDataMapped, padding: 0.5 })
 
       seriesScale = scalePoint({
         domain: config.runtime.barSeriesKeys || config.runtime.seriesKeys,
@@ -94,7 +105,14 @@ const SparkLine: React.FC<SparkLineProps> = props => {
 
   return (
     <ErrorBoundary component='SparkLine'>
-      <svg role='img' aria-label={handleChartAriaLabels(config)} width={parentWidth} height={100} className={'sparkline'} tabIndex={0}>
+      <svg
+        role='img'
+        aria-label={handleChartAriaLabels(config)}
+        width={parentWidth}
+        height={100}
+        className={'sparkline'}
+        tabIndex={0}
+      >
         <title>{`Spark line graphic with the title ${config.title ? config.title : 'No Title Found'}`}</title>
         {config.runtime.lineSeriesKeys?.length > 0
           ? config.runtime.lineSeriesKeys
@@ -104,30 +122,70 @@ const SparkLine: React.FC<SparkLineProps> = props => {
                   style={{ width }}
                   className='sparkline-group'
                   key={`series-${seriesKey}`}
-                  opacity={config.legend.behavior === 'highlight' && seriesHighlight.length > 0 && seriesHighlight.indexOf(seriesKey) === -1 ? 0.5 : 1}
-                  display={config.legend.behavior === 'highlight' || seriesHighlight.length === 0 || seriesHighlight.indexOf(seriesKey) !== -1 ? 'block' : 'none'}
+                  opacity={
+                    config.legend.behavior === 'highlight' &&
+                    seriesHighlight.length > 0 &&
+                    seriesHighlight.indexOf(seriesKey) === -1
+                      ? 0.5
+                      : 1
+                  }
+                  display={
+                    config.legend.behavior === 'highlight' ||
+                    seriesHighlight.length === 0 ||
+                    seriesHighlight.indexOf(seriesKey) !== -1
+                      ? 'block'
+                      : 'none'
+                  }
                 >
-                  {config.labels && data.map((d, dataIndex) => {
-                    return (
-                      <Group key={`series-${seriesKey}-point-${dataIndex}`}>
-                        <Text x={xScale(getXAxisData(d))} y={yScale(getYAxisData(d, seriesKey))} fill={colorScale ? colorScale(config.runtime.seriesLabels ? config.runtime.seriesLabels[seriesKey] : seriesKey) : '#000'} textAnchor='middle'>
-                          {formatNumber(d[seriesKey])}
-                        </Text>
-                      </Group>
-                    )
-                  })}
+                  {config.labels &&
+                    data.map((d, dataIndex) => {
+                      return (
+                        <Group key={`series-${seriesKey}-point-${dataIndex}`}>
+                          <Text
+                            x={xScale(getXAxisData(d, config, parseDate))}
+                            y={yScale(getYAxisData(d, seriesKey))}
+                            fill={
+                              colorScale
+                                ? colorScale(
+                                    config.runtime.seriesLabels ? config.runtime.seriesLabels[seriesKey] : seriesKey
+                                  )
+                                : '#000'
+                            }
+                            textAnchor='middle'
+                          >
+                            {formatNumber(d[seriesKey])}
+                          </Text>
+                        </Group>
+                      )
+                    })}
                   <LinePath
                     curve={allCurves.curveLinear}
                     data={data}
-                    x={d => xScale(getXAxisData(d))}
+                    x={d => xScale(getXAxisData(d, config, parseDate))}
                     y={d => yScale(getYAxisData(d, seriesKey))}
-                    stroke={colorScale ? colorScale(config.runtime.seriesLabels ? config.runtime.seriesLabels[seriesKey] : seriesKey) : '#000'}
+                    stroke={
+                      colorScale
+                        ? colorScale(config.runtime.seriesLabels ? config.runtime.seriesLabels[seriesKey] : seriesKey)
+                        : '#000'
+                    }
                     strokeWidth={2}
                     strokeOpacity={1}
                     shapeRendering='geometricPrecision'
                     markerEnd={`url(#${'arrow'}--${index})`}
                   />
-                  <MarkerArrow id={`arrow--${index}`} refX={2} size={6} markerEnd={`url(#${'arrow'}--${index})`} strokeOpacity={1} fillOpacity={1} fill={colorScale ? colorScale(config.runtime.seriesLabels ? config.runtime.seriesLabels[seriesKey] : seriesKey) : '#000'} />
+                  <MarkerArrow
+                    id={`arrow--${index}`}
+                    refX={2}
+                    size={6}
+                    markerEnd={`url(#${'arrow'}--${index})`}
+                    strokeOpacity={1}
+                    fillOpacity={1}
+                    fill={
+                      colorScale
+                        ? colorScale(config.runtime.seriesLabels ? config.runtime.seriesLabels[seriesKey] : seriesKey)
+                        : '#000'
+                    }
+                  />
                 </Group>
                 <AxisBottom
                   top={yMax + margin.top}
