+++ packages/chart/src/helpers/getColorPalette.js
@@ -1,7 +1,6 @@
 import { colorPalettesChart, twoColorPalette } from '@cdc/core/data/colorPalettes'
-import { useEffect } from 'react'
 
-export const useColorPalette = (config, updateConfig) => {
+export const getColorPalette = config => {
   let twoColorPalettes = []
   let sequential = []
   let nonSequential = []
@@ -39,38 +38,6 @@ export const useColorPalette = (config, updateConfig) => {
     sequential = seqPalettes
     nonSequential = nonSeqPalettes
   }
-
-  // Update pairedBar.palette based on isPaletteReversed
-  useEffect(() => {
-    let palette = ''
-
-    if (config.twoColor.isPaletteReversed && !config.twoColor.palette.endsWith('reverse')) {
-      palette = config.twoColor.palette + 'reverse'
-    }
-
-    if (!config.twoColor.isPaletteReversed && config.twoColor.palette.endsWith('reverse')) {
-      palette = config.twoColor.palette.slice(0, -7)
-    }
-
-    updateConfig({ ...config, twoColor: { ...config.twoColor, palette: palette } })
-  }, [config.twoColor.isPaletteReversed])
-
-  // Update palette based on isPaletteReversed
-  useEffect(() => {
-    let palette = ''
-
-    if (config.isPaletteReversed && !config.palette.endsWith('reverse')) {
-      palette = config.palette + 'reverse'
-    }
-
-    if (!config.isPaletteReversed && config.palette.endsWith('reverse')) {
-      palette = config.palette.slice(0, -7)
-    }
-
-    updateConfig({ ...config, palette: palette })
-  }, [config.isPaletteReversed])
-
-  // Return all palettes
 
   return { twoColorPalettes, sequential, nonSequential, accessibleColors }
 }
