+++ packages/chart/src/hooks/useTooltip.tsx
@@ -14,7 +14,6 @@ export const useTooltip = props => {
     tableData: data,
     config,
     formatNumber,
-    capitalize,
     formatDate,
     formatTooltipsDate,
     parseDate,
@@ -522,7 +521,7 @@ export const useTooltip = props => {
     if (visualizationType === 'Forest Plot') {
       if (key === config.xAxis.dataKey)
         return (
-          <li className='tooltip-heading'>{`${capitalize(config.xAxis.dataKey ? `${config.xAxis.dataKey}: ` : '')} ${
+          <li className='tooltip-heading'>{`${_.capitalize(config.xAxis.dataKey ? `${config.xAxis.dataKey}: ` : '')} ${
             isDateScale(yAxis) ? formatDate(parseDate(key, false)) : value
           }`}</li>
         )
@@ -535,14 +534,14 @@ export const useTooltip = props => {
     // TOOLTIP HEADING
     if (visualizationType === 'Bar' && orientation === 'horizontal' && key === config.xAxis.dataKey)
       return (
-        <li className='tooltip-heading'>{`${capitalize(
+        <li className='tooltip-heading'>{`${_.capitalize(
           config.runtime.yAxis.label ? `${config.runtime.yAxis.label}: ` : ''
         )} ${config.xAxis.type === 'date' ? formattedDate : value}`}</li>
       )
 
     if (key === config.xAxis.dataKey)
       return (
-        <li className='tooltip-heading'>{`${capitalize(
+        <li className='tooltip-heading'>{`${_.capitalize(
           config.runtime.xAxis.label ? `${config.runtime.xAxis.label}: ` : ''
         )} ${isDateScale(xAxis) ? formattedDate : value}`}</li>
       )
