+++ packages/chart/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 
-import CdcChart from './CdcChartWrapper'
+import CdcChart from './CdcChart'
 import './coreStyles_chart.scss'
 
 import '@cdc/core/styles/cove-main.scss'
