+++ packages/chart/src/types/ChartConfig.ts
@@ -79,8 +79,7 @@ type Exclusions = {
 }
 
 export type Legend = CoreLegend & {
-  seriesHighlight: string[]
-
+  seriesHighlight: string[] | []
   hideSuppressionLink: boolean
   style: 'circles' | 'boxes' | 'gradient' | 'lines'
   subStyle: 'linear blocks' | 'smooth'
