+++ packages/chart/src/types/ChartContext.ts
@@ -13,7 +13,6 @@ export type TransformedData = {
 type SharedChartContext = {
   animatedChart?: boolean
   brushConfig: { data: []; isBrushing: boolean; isActive: boolean }
-  capitalize: (value: string) => string
   clean: Function
   colorScale?: ColorScale
   config: ChartConfig
